Moved token generation
